Inline block extraction strategies
